Refactored order handler code to pull out repeated logic
created two functions in common.go that valid and return uuids past in the URL path
and another that marshalls models into json and returns to requester.

attempted to build a middleware for uuid path validation, but doesn't work as of yet.
anytime it pulls the {id} out of path it registers as nothing.

also refactored the routes into groups for easier application of middlewares to specific
routes.